pull request comments fixed
